+++ plonkish_backend/src/backend/unihyperplonk/preprocessor.rs
@@ -1,1 +1,1 @@
-pub(super) use crate::backend::hyperplonk::preprocessor::{batch_size, compose, permutation_polys};
+pub(super) use crate::backend::hyperplonk::preprocessor::{batch_size, preprocess};
